Lesson2_final_part

+++ src/second/lesson/task_8.java
@@ -1,14 +1,16 @@
 package second.lesson;
 
 public class task_8 {
-    public static void main(String[] args){
-    int[] array={1,-10,5,6,45,23,-45,-34,0,32,56,-1,2,-2};
-    int maximum =array[0];                              // максимальное значение
-    int summa=0;                                        //сумма положительных элементов
-    int summa2=0;                                     // сумма отрицательных четных элементов
-    int summa3=0;                                    //  сумма отрицательных элементов
-    int quantity=0;                                  // кол-во положительных элементов
-    int quantity2=0;                                  //кол-во отр. элементов
+    public static void main(String[] args) {
+        int[] array={1,-10,5,6,45,23,-45,-34,0,32,56,-1,2,-2};
+
+        // максимальное значение
+        int maximum =array[0];
+        int summa = 0;                                        //сумма положительных элементов
+        int summa2 = 0;                                     // сумма отрицательных четных элементов
+        int summa3 = 0;                                    //  сумма отрицательных элементов
+        int quantity = 0;                                  // кол-во положительных элементов
+        int quantity2 = 0;                                  //кол-во отр. элементов
         for (int j : array) {
             if (maximum < j) maximum = j;
             if (j > 0) {
@@ -21,12 +23,12 @@ public class task_8 {
                 quantity2++;
             }
         }
-        // средне арифмитическое отр. элементов
-        double srednee = summa3 / quantity2;
-    System.out.println("максимальное значение: "+ maximum);
-    System.out.println("сумма положительных элементов: "+ summa);
-    System.out.println("сумма четных отрицательных элементов: "+ summa2);
-    System.out.println("количество положительных элементов: "+ quantity);
-    System.out.println("среднее арифметическое отрицательных элементов: "+ srednee);
+
+        double srednee = (double)summa3 / (double)quantity2;                    // средне арифмитическое отр. элементов
+        System.out.println("максимальное значение: "+ maximum);
+        System.out.println("сумма положительных элементов: "+ summa);
+        System.out.println("сумма четных отрицательных элементов: "+ summa2);
+        System.out.println("количество положительных элементов: "+ quantity);
+        System.out.println("среднее арифметическое отрицательных элементов: "+ srednee);
     }
 }
